Add initial support for using git CLI when available
JGit has some feature holes/defects -- in particular, it lacks support
for git work-trees and has so for a long time.

This adds initial support for using the git CLI when it is available,
with support for falling back to JGit when not. This means features
like work-trees are recognized and work properly.

Changelog generation has not yet been moved to using GitProvider, as it
requires more work to detach it from its use of JGit classes.