Fix configuration cache issues in buildscript
The two ad-hoc tasks are changed into proper task classes, for proper
task property support (and the use of ExecOperations in the changelog
task).

Unfortunately, the licenser plugin's tasks cause problems when reusing a
configuration cache entry that includes those tasks, such as running the
`build` task.

Current workaround is to avoid license checks when configuration cache
is enabled, such as running `assemble` instead of `build`.