Use value source to retrieve git info
JGit executes the `git` command to determine some information about the
system git config. However, executing commands via Java's built-in
`ProcessBuilder` is unsupported in Gradle when the configuration cache
is enabled.

Gradle does offer us an out, though: calculations in a `ValueSource` are
done 'outside' of the confines of the configuration cache, and are
therefore free to e.g., use `ProcessBuilder` instead of Gradle's own
process execution facilities (`ExecOperations`).

This has a minor cost: the value source is always evaluated during the
configuration phase, and a change in its reported value causes a
configuration cache invalidation. However, we are fine with that cost.