Refactor utility classes to be package-private
This gives us more flexibility to change around the internals of
GradleUtils in the future.

To resolve the problem of the subpackages needing access to certain
methods in the main package, the `InternalAccessor` class exists for
those specific methods. The class is specifically not part of the API,
and may change or be removed at any time.