Add new versioning API
The configuration for the version is now unified in the extension, with
only one method to actually calculate the version.

The new version calculation logic is properly hidden behind
package-private classes, to allow future work on improving and
refactoring these classes without affecting the API (such as for `git`
CLI support).

This separation of concerns also means that the changelog generation
system can use the same version calculation logic, to have matching
versions between changelog and the actual project (published artifacts).

Currently, the new API is exposed as a separate `ngu` extension on the
project, to allow concurrent use of both APIs for comparison testing.
Once the new API is finalized, it will subsume the old API.