Workaround issues in changelog task inputs
Because Gradle keeps track of inputs to tasks, and the changelog task
has the project directory as its input, it tries to track the *whole*
project directory for the task, which leads to various task-input-output
errors (Gradle thinking the changelog task depends on the output of the
Jar task, for example).

We circumvent for now by specifically targeting the `.git` subdirectory,
on the premise that the plugin is likely to be applied on the root
project at the repository root.

This will be properly fixed at a later date, once an acceptable solution
can be found.