Remove upwards traversal for changelog task
Changelog publishing only works for the changelog of that specific
project. However, because Git traverses upwards for the repository,
this should have no adverse effect for e.g. a subproject publishing o
behalf of the whole project.

This also cleans up some code to use the task configuration avoidance
API.